Change table name to status_logs
